Added script stub for quest 9438 (todo with CreatureScript update coming soon), removed logging into non existent table character_purchases
Former-commit-id: 5339f9171c0ee5713aa8984d8bc79d38c041250d

+++ src/game/Level3.cpp
@@ -2613,7 +2613,7 @@ bool ChatHandler::HandleListCreatureCommand(const char* args)
     result=WorldDatabase.PQuery("SELECT COUNT(guid) FROM creature WHERE id='%u'",cr_id);
     if(result)
     {
-        cr_count = (*result)[0].GetUInt32();
+        cr_count = (*result)[0].GetUInt64();
     }
 
     if(m_session)
